+++ src/bot.ts
@@ -1,18 +1,19 @@
-import { Client, Collection as DiscordCollection } from 'discord.js'
+import { Client } from 'discord.js'
 import { loadCommands, loadEvents, loadDashboard } from './utils/loader';
 import Command from './interfaces/command';
 import Event from './interfaces/event';
 import { MongoClient, Collection, Db as Database } from 'mongodb';
+import { User, Server, Permission, Tag } from './interfaces/databaseStructures';
 
 export default class Bot {
     client: Client = new Client({ disableEveryone: true });
     commands: Command[] = [];
     events: Event[] = [];
 
-    users: Collection;
-    servers: Collection;
-    permissions: Collection;
-    tags: Collection;
+    users: Collection<User>;
+    servers: Collection<Server>;
+    permissions: Collection<Permission>;
+    tags: Collection<Tag>;
 
     music: any = {};
 
@@ -21,7 +22,7 @@ export default class Bot {
 
         let conn: MongoClient = await MongoClient.connect('mongodb://localhost:27017', { useNewUrlParser: true });
         let database: Database = conn.db('izel');
-    
+
         this.users = database.collection('users');
         this.servers = database.collection('servers');
         this.permissions = database.collection('permissions');