+++ src/interfaces/permission.ts
@@ -1,4 +1,0 @@
-export default interface Permission {
-    action: string;
-    user_ids: string[];
-}